+++ netring/examples/l7/full_monitor.rs
@@ -121,8 +121,8 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
                     totals.http_req += 1;
                     println!(
                         "[HTTP] →  {method} {path} {ver:?}",
-                        method = req.method,
-                        path = req.path,
+                        method = String::from_utf8_lossy(&req.method),
+                        path = String::from_utf8_lossy(&req.path),
                         ver = req.version
                     );
                 }
@@ -131,7 +131,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
                     println!(
                         "[HTTP] ←  {status} {reason}  {len} bytes",
                         status = resp.status,
-                        reason = resp.reason,
+                        reason = String::from_utf8_lossy(&resp.reason),
                         len = resp.body.len()
                     );
                 }